no test grad

+++ Tfunction.h
@@ -14,9 +14,9 @@ class TFunction : public std::enable_shared_from_this<TFunction>{
 public:
     virtual ~TFunction() = default;
 
-    virtual double operator()(double x) const = 0;
+    virtual long double operator()(long double x) const = 0;
 
-    virtual double GetDeriv(double x) const = 0;
+    virtual long double GetDeriv(long double x) const = 0;
 
     virtual std::string ToString() const = 0;
 
@@ -28,60 +28,60 @@ public:
 
 class IdentityFunction : public TFunction {
 public:
-    double operator()(double x) const override;
+    long double operator()(long double x) const override;
 
-    double GetDeriv(double x) const override;
+    long double GetDeriv(long double x) const override;
 
     std::string ToString() const override;
 };
 
 class ConstantFunction : public TFunction {
 private:
-    double value_;
+    long double value_;
 
 public:
-    explicit ConstantFunction(double v);
+    explicit ConstantFunction(long double v);
 
-    double operator()(double x) const override;
+    long double operator()(long double x) const override;
 
-    double GetDeriv(double x) const override;
+    long double GetDeriv(long double x) const override;
 
     std::string ToString() const override;
 };
 
 class PowerFunction : public TFunction {
 private:
-    double exponent_;
+    long double exponent_;
 
 public:
-    explicit PowerFunction(double exp);
+    explicit PowerFunction(long double exp);
 
-    double operator()(double x) const override;
+    long double operator()(long double x) const override;
 
-    double GetDeriv(double x) const override;
+    long double GetDeriv(long double x) const override;
 
     std::string ToString() const override;
 };
 
 class ExpFunction : public TFunction {
 public:
-    double operator()(double x) const override;
+    long double operator()(long double x) const override;
 
-    double GetDeriv(double x) const override;
+    long double GetDeriv(long double x) const override;
 
     std::string ToString() const override;
 };
 
 class PolynomialFunction : public TFunction {
 private:
-    std::vector<double> coeff_;
+    std::vector<long double> coeff_;
 
 public:
-    explicit PolynomialFunction(const std::vector<double>& coeffs);
+    explicit PolynomialFunction(const std::vector<long double>& coeffs);
 
-    double operator()(double x) const override;
+    long double operator()(long double x) const override;
 
-    double GetDeriv(double x) const override;
+    long double GetDeriv(long double x) const override;
 
     std::string ToString() const override;
 };
@@ -95,9 +95,9 @@ private:
 public:
     explicit Function(std::shared_ptr<const TFunction> f, std::shared_ptr<const TFunction> s, char op);
 
-    double operator()(double x) const override;
+    long double operator()(long double x) const override;
 
-    double GetDeriv(double x) const override;
+    long double GetDeriv(long double x) const override;
 
     std::string ToString() const override;
 };
